+++ grid2op/Agent/oneChangeThenNothing.py
@@ -6,7 +6,7 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 
-from grid2op.Agent.BaseAgent import BaseAgent
+from grid2op.Agent.baseAgent import BaseAgent
 
 
 class OneChangeThenNothing(BaseAgent):