+++ grid2op/Agent/recoPowerlineAgent.py
@@ -6,7 +6,7 @@
 # SPDX-License-Identifier: MPL-2.0
 # This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
 import numpy as np
-from grid2op.Agent.GreedyAgent import GreedyAgent
+from grid2op.Agent.greedyAgent import GreedyAgent
 
 
 class RecoPowerlineAgent(GreedyAgent):